ChatBotGenerativo _SoulCode/requirements.txt
gradio>=4.0.0
langchain>=0.1.13
langchain-openai>=0.0.8
openai>=1.0.0
python-dotenv>=1.0.0